Add links to some of my things
When people come to websites they usually to find some sort of content.
As this is a personal page the content is me, so this adds some a couple
links to other things of mine.

This also changes the way positioning is done in CSS, it feels less
hacky so I guess that's good.